+++ SRML/Drivers/Devices/MPU6050/mpu6050_config.h
@@ -56,7 +56,7 @@ extern IIC_PIN_Typedef MPU6050_IIC_PIN;
 extern MPUData_Typedef MPUData;
 
 /* Exported function declarations --------------------------------------------*/
-uint8_t MPU6050_Init(GPIO_TypeDef *gpiox,uint32_t scl_pinx,uint32_t sda_pinx);
+uint8_t MPU6050_Init(GPIO_TypeDef *gpiox,uint16_t scl_pinx,uint16_t sda_pinx);
 #ifdef __cplusplus
 }
 #endif
